chore: rename all camel case files to kebab case
