+++ src/operation/select-rule.ts
